Chore: modify environment variables - fixed layout error
