+++ src/components/MainLayout.tsx
@@ -1,0 +1,25 @@
+import { Geist, Geist_Mono } from 'next/font/google';
+
+const geistSans = Geist({
+  variable: '--font-geist-sans',
+  subsets: ['latin'],
+});
+
+const geistMono = Geist_Mono({
+  variable: '--font-geist-mono',
+  subsets: ['latin'],
+});
+
+export default function MainLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <div
+      className={`${geistSans.variable} ${geistMono.variable} antialiased min-w-full min-h-screen flex justify-center items-center`}
+    >
+      {children}
+    </div>
+  );
+}
